fix: use even id to resolve no-unused variavble error

+++ app/api/webhooks/clerk/route.ts
@@ -52,10 +52,10 @@ export async function POST(req: Request) {
   }
 
   // Get the ID and type
-  // const { id } = evt.data;
+  const { id } = evt.data;
   const eventType = evt.type;
 
-  if (eventType === 'user.created') {
+  if (id && eventType === 'user.created') {
     const { id, email_addresses, image_url, first_name, last_name, username } =
       evt.data;
 
@@ -82,7 +82,7 @@ export async function POST(req: Request) {
     return NextResponse.json({ message: 'OK', user: newUser });
   }
 
-  if (eventType === 'user.updated') {
+  if (id && eventType === 'user.updated') {
     const { id, image_url, first_name, last_name, username } = evt.data;
 
     const user = {
@@ -97,7 +97,7 @@ export async function POST(req: Request) {
     return NextResponse.json({ message: 'OK', user: updatedUser });
   }
 
-  if (eventType === 'user.deleted') {
+  if ( id && eventType === 'user.deleted') {
     const { id } = evt.data;
 
     const deletedUser = await deleteUser(id!);
